BAU: Move non unit test files to utils
The files containing no unit tests are moved to utils to enable
developers to find unit tests easily.

Authors: @adityapahuja
