+++ src/test/java/uk/gov/ida/eventemitter/EventEmitterIntegrationTest.java
@@ -21,6 +21,11 @@ import org.junit.AfterClass;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.junit.runner.RunWith;
+import uk.gov.ida.eventemitter.utils.AmazonHelper;
+import uk.gov.ida.eventemitter.utils.TestConfiguration;
+import uk.gov.ida.eventemitter.utils.TestDecrypter;
+import uk.gov.ida.eventemitter.utils.TestEvent;
+import uk.gov.ida.eventemitter.utils.TestEventEmitterModule;
 
 import java.nio.ByteBuffer;
 import java.util.HashMap;
