+++ src/test/java/uk/gov/ida/eventemitter/EventEmitterTest.java
@@ -7,6 +7,7 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.Mock;
 import org.mockito.runners.MockitoJUnitRunner;
+import uk.gov.ida.eventemitter.utils.TestEvent;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
